Handle non-UTF-8 rfc822_mail correctly
Emit the unparsed RFC822 mail as the event message converted to UTF-8
using the `plain` codec and as the field `[@metadata][rfc822_mail]`
in the original `ASCII-8BIT` encoding. Whenever non-UTF-8 mail is
converted to UTF-8, 8-bit characters are corrupted. Therefore, the
`[@metadata][rfc822_mail]` field must be used for further processing
in the filters.